Update pull request description instead of leaving a comment
Let's update the pull request description instead
since leaving a comment does not trigger a pull request
workflow. A workaround could be to make a dummy change to
the pull request description from an issue_comment workflow
however that would require a Personal Access Token to be
used for no good reason, which makes things more complex
and potentially insecure.